% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_helpers.R
\name{arc.strength.dbn}
\alias{arc.strength.dbn}
\title{Calculate arc strengths from dbn object.}
\usage{
arc.strength.dbn(x, data)
}
\arguments{
\item{x}{dbn onbject.  Usually output of \code{dbn_learn()}.}

\item{data}{Unlagged data used to learn structure.}
}
\value{
Dataframe with four columns:
\itemize{
\item from: Parent node
\item to: Child node
\item lag: Markov lag of connection
\item strength: Strength of this connection
}
}
\description{
Calculate arc strengths from dbn object.
}
