% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_helpers.R
\name{as.adjacency.dbn}
\alias{as.adjacency.dbn}
\title{Get adjacency matrix of rolled dbn structure}
\usage{
as.adjacency.dbn(x)
}
\arguments{
\item{x}{dbn object. Usually output of \code{dbn_learn()}.}
}
\value{
Adjacency matrix showing connectivity of dbn object with from nodes in columns and to nodes in rows.
Note that by definition self-connections are present.
}
\description{
Get adjacency matrix of rolled dbn structure
}
