% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_learn.R
\name{dbn_learn}
\alias{dbn_learn}
\title{Learn dynamic Bayesian network structure.}
\usage{
dbn_learn(data, score, minMarkovLag = 1, maxMarkovLag = 1, search = "hc", ...)
}
\arguments{
\item{data}{Data frame with variables in column and observations in rows.}

\item{score}{Score to be used implemented using \pkg{bnlearn}.}

\item{minMarkovLag}{Minimum markov lag to be used in analysis. Must be less than maximum Markov lag and ≥ 0. If 0, output is not a true DBN as it may contain connections within time slice.}

\item{maxMarkovLag}{Maximum Markov lag to be used in analysis. Minimum value of 1.}

\item{search}{Search type implemented using \pkg{bnlearn}.
Must be one of:
\itemize{
\item \code{"hc"}: hill-climbing search
\item \code{"tabu"}: tabu search
}}

\item{...}{Other arguments to be supplied to \code{\link[bnlearn]{hc}} or \code{\link[bnlearn]{tabu}}.}
}
\description{
Learn dynamic Bayesian network structure.
}
