% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dynamic_data.R
\name{make_dynamic_data}
\alias{make_dynamic_data}
\title{Takes dataframe containing time-series observations of variables and reformats as data for DBN structure learning,
where each column corresponds to a node in the dynamic network.}
\usage{
make_dynamic_data(data, minMarkovLag = 1, maxMarkovLag = 1)
}
\arguments{
\item{data}{Data frame containing observations in rows and variables in columns.}

\item{minMarkovLag}{Minimum Markov lag to be used in analysis. Must be less than maximum Markov lag and ≥ 0.
Value of 0 is allowed experimentally and output may include  within time-slice connections.
This is not recommended and should only be considered where information lag in system is less than the temporal resolution of data.}

\item{maxMarkovLag}{Maximum Markov lag to be used in analysis. Minimum value of 1.}
}
\value{
Dataframe containing lagged columns of all variables up to maxMarkovLag.
}
\description{
Takes dataframe containing time-series observations of variables and reformats as data for DBN structure learning,
where each column corresponds to a node in the dynamic network.
}
