% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wb_lists.R
\name{make_wb_lists}
\alias{make_wb_lists}
\title{Makes Whitelist and Blacklist for enforcing basic DBN structure}
\usage{
make_wb_lists(
  varNames = NULL,
  nNodes = NULL,
  minMarkovLag = 1,
  maxMarkovLag = 1
)
}
\arguments{
\item{varNames}{Names of variables in the network. These should match column names of data frame passed to make_data.}

\item{nNodes}{Number of nodes in static network (ie number of variables in analysis). It is only recommended to set this parameter if not supplying variable names, in this case nNodes variable names will be generated in the form Ni..}

\item{minMarkovLag}{The minimum time lag in the DBN structure.}

\item{maxMarkovLag}{The maximum time lag in the DBN structure.}
}
\value{
list  with whitelist in first element and blacklist in second.
}
\description{
This function creates a list containing 2 vectors. The first contains links which should be whitelisted
and the second contains links which should be blacklisted when learning dynamic BN structures.
Note that although the maximum markov lag is taken as an input, minimum markov lag will always be 1.
Uses same node naming convention (NODE_t) as \code{make_dynamic_data()}.
}
