% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn-class.R
\name{plot.dbn}
\alias{plot.dbn}
\title{Plots dbn structure.}
\usage{
\method{plot}{dbn}(
  x,
  roll = T,
  whitelist = T,
  ylim = c(0, 1200),
  xlim = ylim,
  radius = 300,
  arrow = 80,
  cex = 8
)
}
\arguments{
\item{x}{dbn project to be plotted. Will usually be output of \code{dbn_learn()}.}

\item{roll}{Plot rolled form of dbn?}

\item{whitelist}{If plotting unrolled dbn, should whitelist links be shown. If true these links will be shown in grey.}

\item{ylim}{Y coordinate limits for plotting window. Can be used to resize plotting pane. In form \code{c(lowerLimit,upperLimit)}.}

\item{xlim}{X coordinate limits for plotting window. Can be used to resize plotting pane. In form \code{c(lowerLimit,upperLimit)}.}

\item{radius}{When plotting the rolled network, the radius of circle around which nodes will be arranged.}

\item{arrow}{Offset from centre of node to arrow-head. May have to be fine tuned depending on number of nodes in network.}

\item{cex}{Controls node size.}
}
\description{
Plots dbn structure.
}
