% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn-class.R
\name{print.dbn}
\alias{print.dbn}
\title{Prints Information on rolled structure, followed by standard unlearn information on unrolled structure}
\usage{
\method{print}{dbn}(x)
}
\arguments{
\item{x}{Object of dbn class}
}
\description{
Prints Information on rolled structure, followed by standard unlearn information on unrolled structure
}
\details{
This is the S3 \code{print()} method for objects of class \code{dbn}. Prints basic information on the dbn.
Note to access bnlearns print output for information on the unrolled model use \code{print(x$bn)}.
}
