% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_dbn.R
\name{roll_dbn}
\alias{roll_dbn}
\title{Takes a DBN structure learned using bn learn (bn object) and outputs a dbn object}
\usage{
roll_dbn(bn, minMarkovLag = NULL)
}
\arguments{
\item{bn}{The bn object representing the DBN. Node names must be in the format \{VARIABLENAME\}_t-\{LAG\} for lagged cols and \{VARIABLENAME\}_t for time slice 0.}

\item{minMarkovLag}{This parameter is included for the edge case where within slice connections were allowed (minMarkovLag =0) but none were learned.
If not supplied, it will default to lowest lag a connection came frome, which will be one for this case, for all other cases this will give the correct value.}
}
\value{
dbn object were first element contains information on rolled dbn structure (learning process, connections in rolled network, nodes),
and the second element contains the unrolled bn representation.
}
\description{
Takes a DBN structure learned using bn learn (bn object) and outputs a dbn object
}
